+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [NEXT]
 
 - add benchmark result
+- re-export Cow, Owned, and Borrowed for different COW implementations
 
 ## [0.1.1] - 2023-09-06